Time: 0 ms (100%), Space: 19.8 MB (44.88%) - LeetHub

+++ 0662-maximum-width-of-binary-tree/0662-maximum-width-of-binary-tree.cpp
@@ -29,7 +29,7 @@ public:
                 if(node->left) q.push({node->left,cur_idx*2+1});
                 if(node->right) q.push({node->right,cur_idx*2+2});
             }
-            ans=max(ans,last-first+1);
+            ans=max(ans,static_cast<int>(last - first + 1));
         }
         return ans;
     }
